Added ability to conditionally replace things based on enabled patches.
A list of patches can be specified for Replaced, and the replacement is only done if at least one of the listed patches is enabled. Inactive Replaced methods still act as decoys, so they can be called from other, active Replaced methods, but the method will keep its unpatched behavior.