Switch back to 5 checks per cycle
We can bypass the check that used to be the first because the newly added checks should already catch anything it could catch. This gives us 5 checks per cycle again, and a negligible performance loss compared to the vanilla game. The cost is a bit of ugliness at the collision check call.